refactor(core): split dpi/quic.rs into a quic/ directory module
