+++ crates/rustnet-core/src/network/dpi/quic/mod.rs
@@ -1,0 +1,10 @@
+mod crypto;
+mod packet;
+mod salvage;
+mod tls;
+
+#[cfg(test)]
+mod tests;
+
+pub use packet::{is_quic_packet, parse_quic_packet};
+pub use tls::{is_partial_sni, try_extract_tls_from_reassembler};
